delete mail data

+++ modules/Apps/MailClient/MailClient_0.php
@@ -322,7 +322,7 @@ class Apps_MailClient extends Module {
 			$r->add_data($row['mail']);
 			$r->add_action($this->create_callback_href(array($this,'account'),array($row['id'],'edit')),'Edit');
 //			$r->add_action($this->create_callback_href(array($this,'account'),array($row['id'],'view')),'View');
-			$r->add_action($this->create_confirm_callback_href($this->lang->ht('Are you sure?'),array($this,'delete_account'),$row['id']),'Delete');
+			$r->add_action($this->create_callback_href(array($this,'delete_account'),$row['id']),'Delete');
 		}
 		$this->display_module($gb);
 		Base_ActionBarCommon::add('add','New account',$this->create_callback_href(array($this,'account'),array(null,'new')));
@@ -422,7 +422,29 @@ class Apps_MailClient extends Module {
 	}
 
 	public function delete_account($id){
-		DB::Execute('DELETE FROM apps_mailclient_accounts WHERE id=%d',array($id));
+		if($this->is_back()) return false;
+		$f = $this->init_module('Libs/QuickForm');
+		$mail = DB::GetOne('SELECT mail FROM apps_mailclient_accounts WHERE id=%d',array($id));
+		$f->addElement('header',null,$this->lang->t('Delete account %s',array($mail)));
+		$f->addElement('checkbox','setup',$this->lang->t('Account setup'));
+		$f->addElement('checkbox','data',$this->lang->t('Emails'));
+		$f->setDefaults(array('setup'=>1));
+
+		if($f->validate()) {
+			$values = $f->exportValues();
+			if(isset($values['setup']) && $values['setup'])
+				DB::Execute('DELETE FROM apps_mailclient_accounts WHERE id=%d',array($id));
+			if(isset($values['data']) && $values['data'])
+				recursive_rmdir($this->get_data_dir().Acl::get_user().'/'.str_replace(array('@','.'),array('__at__','__dot__'),$mail));
+			return false;
+		}
+		$f->display();
+
+		Base_ActionBarCommon::add('back','Back',$this->create_back_href());
+		Base_ActionBarCommon::add('delete','Delete',' href="javascript:void(0)" onClick="'.addcslashes($f->get_submit_form_js(),'"').'"');
+		
+		
+		return true;
 	}
